+++ src/onevision/cv/io/video.py
@@ -141,7 +141,7 @@ class VideoLoader:
 				
 				self.index += 1
 
-			return np.array(images, dtype=np.uint8), indexes, files, rel_paths
+			return np.array(images), indexes, files, rel_paths
 
 	def __del__(self):
 		"""Close the `video_capture` object."""
